ビルドエラー対応：authOptionsを外部ファイルに切り出し
